backend/ocr/requirements.txt
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.0
Pillow==10.1.0
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic==2.4.2
scikit-image==0.22.0
